# services/api-gateway-go/internal/http/openapi.yaml
openapi: 3.1.0
info:
  title: Trustworthy Stock Intelligence API
  version: 0.1.0
  summary: PostgreSQL-backed warning, watchlist, and ticker analysis API.
  description: |
    The API serves precomputed stock drawdown-risk warning records generated
    by the Python ML core and stored in PostgreSQL. Request handlers do not run
    synchronous inference.
servers:
  - url: http://127.0.0.1:18080
    description: Local development API
...
paths:
...
  /api/v1/watchlists/{name}/tickers:
    post:
      tags: [Watchlists]
      summary: Add a ticker to a watchlist.
      operationId: addWatchlistTicker
      parameters:
        - $ref: "#/components/parameters/WatchlistNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistAddTickerRequest"
      responses:
        "201":
          description: Updated watchlist state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/v1/watchlists/{name}/tickers/{ticker}:
    delete:
      tags: [Watchlists]
      summary: Remove a ticker from a watchlist.
      operationId: removeWatchlistTicker
      parameters:
        - $ref: "#/components/parameters/WatchlistNamePath"
        - $ref: "#/components/parameters/TickerPath"
      responses:
        "200":
          description: Updated watchlist state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
    TickerPath:
      name: ticker
      in: path
      required: true
      description: Alphabetic ticker or numeric Taiwan stock code in the loaded batch.
      schema:
        type: string
        minLength: 1
        examples: ["NVDA", "2330"]
    WatchlistNamePath:
      name: name
      in: path
      required: true
      description: Watchlist name. The local dashboard uses `default`.
      schema:
        type: string
        minLength: 1
        examples: ["default"]
  responses:
    BadRequest:
      description: Invalid request parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Ticker was not found in the latest loaded warning batch.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Required backing store is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    WarningLevel:
      type: string
      enum: [alert, watch, abstain, no_alert]
...
    WatchlistAddTickerRequest:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - ticker
        - market
        - notes
      properties:
        schema_version:
          type: string
          const: watchlist_add.v1
        ticker:
          type: string
          examples: ["NVDA", "2330"]
        market:
          type: string
          enum: [auto, us, twse, tpex]
        notes:
          type: string
    WatchlistResponse:
      type: object
      additionalProperties: false
      required:
        - schema_version
        - name
        - record_count
        - updated_at
        - tickers
      properties:
        schema_version:
          type: string
          const: watchlist.v1
        name:
          type: string
        record_count:
          type: integer
          minimum: 0
        updated_at:
          type: string
        tickers:
          type: array
          items:
            $ref: "#/components/schemas/WatchlistTicker"
    WatchlistTicker:
      type: object
      additionalProperties: false
      required:
        - ticker
        - query_symbol
        - market
        - added_at
        - notes
        - has_latest_warning
      properties:
        ticker:
          type: string
        query_symbol:
          type: string
        market:
          type: string
          enum: [us, twse, tpex, taiwan, unknown]
        added_at:
          type: string
        notes:
          type: string
        has_latest_warning:
          type: boolean
        latest_warning:
          $ref: "#/components/schemas/WatchlistLatestWarning"
    WatchlistLatestWarning:
      type: object
      additionalProperties: false
      required:
        - date
        - warning_level
        - calibrated_risk_probability
        - trust_score
      properties:
        date:
          type: string
        warning_level:
          $ref: "#/components/schemas/WarningLevel"
        calibrated_risk_probability:
          type: number
          minimum: 0
          maximum: 1
        trust_score:
          type: number
          minimum: 0
          maximum: 1
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string